[reassoc] don't sign-extend bitwise_fold result
bitwise_fold built `BigInt((int64_t)op(la, lb))`, which sign-extends
any high bit set in the 64-bit result. For >64-bit destination types
(e.g. unsignedbv 128, _ExtInt(N) with N > 64), from_integer's
two's-complement round-trip then extends those bits all the way to
the destination width. e.g.

  uint128_const(1<<63) & uint128_const(1<<63)

would fold to 2^128 - 2^63 instead of 2^63.

Build the BigInt as `unsigned long long` so the high bit doesn't
flip sign. Operands wider than 64 bits still bail via the existing
fits() check, so the >64-bit destination case only matters when the
fold input fits in 64 bits but the output type is wider — which is
the case here.

Codex adversarial-review iteration 11.

+++ src/util/expr_reassociate.cpp
@@ -581,7 +581,12 @@ std::optional<BigInt> bitwise_fold(const BigInt &a, const BigInt &b, U64Op op)
     return std::nullopt;
   uint64_t la = a.is_uint64() ? a.to_uint64() : (uint64_t)a.to_int64();
   uint64_t lb = b.is_uint64() ? b.to_uint64() : (uint64_t)b.to_int64();
-  return BigInt((int64_t)op(la, lb));
+  // Build the BigInt as unsigned. Casting through int64_t would
+  // sign-extend any high bit set in the result, and from_integer() for
+  // wider destination types (e.g. >64-bit unsignedbv) would then sign-
+  // extend further. e.g. (uint128)(1<<63) & (uint128)(1<<63) must stay
+  // 2^63, not become 2^128 - 2^63.
+  return BigInt((unsigned long long)op(la, lb));
 }
 
 void linearize_bitand(
